refactored the server
